requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.3